requirements.txt

# ============================================================
# MCP SDK from GitHub (not on PyPI)
# ============================================================
#git+https://github.com/modelcontextprotocol/python-sdk.git
#mcp
fastmcp
mistralai
python-dotenv
aiohttp
# ============================================================
# LLM OLLAMA
# ============================================================
#ollama>=0.1.9


# ============================================================
# Notes
# ============================================================
# 1. Ensure Python 3.10+ is installed via pyenv or system Python.
# 2. Activate a virtual environment before installing:
#      python3 -m venv .venv
#      source .venv/bin/activate
# 3. Install requirements:
#      pip install -r requirements.txt
